// packages/code/src/views/repo/pull-request/details/pull-request-details-types.ts
export interface DefaultReviewersApprovalsData {
  current_count?: number
  evaluations?: TypesReviewerEvaluation[] | null
  minimum_required_count?: number
  minimum_required_count_latest?: number
  principals?: TypesPrincipalInfo[] | null
  user_groups?: TypesUserGroupInfo[] | null
}
// ...
export interface TypesPullReqActivity {
  author?: TypesPrincipalInfo
  code_comment?: TypesCodeCommentFields
  created?: number
  deleted?: number | null
  edited?: number
  id?: number
  kind?: EnumPullReqActivityKind
  mentions?: {
    [key: string]: TypesPrincipalInfo
  }
  metadata?: TypesPullReqActivityMetadata
  order?: number
  parent_id?: number | null
  payload?: GeneralPayload | TypesPullReqActivity
  pullreq_id?: number
  repo_id?: number
  resolved?: number | null
  resolver?: TypesPrincipalInfo
  sub_order?: number
  text?: string
  type?: EnumPullReqActivityType
  updated?: number
  user_group_mentions?: { [key: string]: TypesUserGroupInfo }
}

export type EnumPrincipalType = 'service' | 'serviceaccount' | 'user'

export interface TypesPrincipalInfo {
  created?: number
  display_name?: string
  email?: string
  id?: number
  type?: EnumPrincipalType
  uid?: string
  updated?: number
}

export interface TypesUserGroupInfo {
  description?: string
  id?: number
  identifier?: string
  name?: string
  scope?: number
}

export declare type EnumPullReqReviewerType = 'assigned' | 'requested' | 'self_assigned' | 'default' | 'code_owners'

export declare type ReviewerListPullReqOkResponse = TypesPullReqReviewer[]
export interface TypesPullReqReviewer {
  added_by?: Partial<PrincipalType>
  created?: number
  latest_review_id?: number | null
  review_decision?: EnumPullReqReviewDecision
  reviewer?: Partial<PrincipalType>
  sha?: string
  type?: EnumPullReqReviewerType
  updated?: number
}

export interface TypesReviewerEvaluation {
  decision?: EnumPullReqReviewDecision
  reviewer?: TypesPrincipalInfo
  sha?: string
}

export interface TypesCodeCommentFields {
  line_new?: number
  line_old?: number
  merge_base_sha?: string
  outdated?: boolean
  path?: string
  source_sha?: string
  span_new?: number
  span_old?: number
}

export type EnumPullReqActivityKind = 'change-comment' | 'comment' | 'system'

export interface TypesPullReqActivityMetadata {
  mentions?: TypesPullReqActivityMentionsMetadata
  suggestions?: TypesPullReqActivitySuggestionsMetadata
}

export interface TypesPullReqActivityMentionsMetadata {
  ids?: number[]
  user_group_ids?: number[]
}
export interface TypesPullReqActivitySuggestionsMetadata {
  applied_check_sum?: string
  applied_commit_sha?: string
  check_sums?: string[]
}

export interface GeneralPayload extends TypesPullReqActivity {
  text?: string
  [key: string]: unknown
  payload?: GeneralPayload
  type?: EnumPullReqActivityType
  kind?: EnumPullReqActivityKind
  message?: string
  reviewer_type?: ReviewerAddActivity
}

export type EnumPullReqActivityType =
  | 'branch-delete'
  | 'branch-update'
  | 'code-comment'
  | 'comment'
  | 'merge'
  | 'review-submit'
  | 'reviewer-delete'
  | 'state-change'
  | 'title-change'
  | 'reviewer-add'
  | 'label-modify'
  | 'branch-restore'
  | 'target-branch-change'
  | 'title-change'
  | 'user-group-reviewer-add'
  | 'user-group-reviewer-delete'

export enum ReviewerAddActivity {
  REQUESTED = 'requested',
  ASSIGNED = 'assigned',
  SELF_ASSIGNED = 'self_assigned',
  DEFAULT = 'default',
  CODEOWNERS = 'code_owners'
}
// ...
export interface TypesViolation {
  code?: string
  message?: string
  params?: unknown
}

export interface TypesRuleViolations {
  bypassable?: boolean
  bypassed?: boolean
  rule?: TypesRuleInfo
  violations?: TypesViolation[] | null
}

export interface TypesRuleInfo {
  identifier?: string
  repo_path?: string
  space_path?: string
  state?: EnumRuleState
  type?: string
}

export type EnumRuleState = 'active' | 'disabled' | 'monitor' | null
// ...
export interface CommentItem<T = unknown> {
  id: number
  author: string
  created: string | number
  edited: string | number
  updated: string | number
  deleted: string | number
  outdated: boolean
  content: string
  payload?: T // optional payload for callers to handle on callback calls
  appliedCheckSum?: string
  checkSums?: string[]
  codeBlockContent?: string
  appliedCommitSha?: string
}

export interface DiffFileEntry extends DiffFile {
  filePath: string
  containerId: string
  contentId: string
  fileViews?: Map<string, string>
  isRename?: boolean
}

export interface DiffFileName {
  oldName: string
  newName: string
}
export interface DiffFile extends DiffFileName {
  addedLines: number
  deletedLines: number
  isCombined: boolean
  isGitDiff: boolean
  language: string
  blocks: DiffBlock[]
  oldMode?: string | string[]
  newMode?: string
  deletedFileMode?: string
  newFileMode?: string
  isDeleted?: boolean
  isNew?: boolean
  isCopy?: boolean
  isRename?: boolean
  isBinary?: boolean
  isTooBig?: boolean
  unchangedPercentage?: number
  changedPercentage?: number
  checksumBefore?: string | string[]
  checksumAfter?: string
  mode?: string
  raw?: string
}
// ...
export interface ApprovalItem {
  id: number
  state?: string
  method: string
  title: string
  items?: ApprovalItems[]
}
// ...
export interface ApprovalItems {
  items: ApprovalItem[]
}
export type ButtonEnum = 'success' | 'muted' | 'default' | 'error' | 'warning' | null | undefined
export type EnumPullReqReviewDecisionExtended = EnumPullReqReviewDecision | 'outdated'
// ...
export interface DiffHeaderProps {
  text: string
  data?: string
  title: string
  lang: string
  addedLines?: number
  deletedLines?: number
  isBinary?: boolean
  isDeleted?: boolean
  unchangedPercentage?: number
  filePath?: string
  fileViews?: Map<string, string>
  checksumAfter?: string
  diffData?: DiffFileEntry
}